refactor: separate snip and speed components
