+++ components/speed/speed-range.js
